Struct rustup_dist::component::Transaction

pub struct Transaction<'a> {
    // some fields omitted
}

A Transaction tracks changes to the file system, allowing them to be rolled back in case of an error. Instead of deleting or overwriting file, the old copies are moved to a temporary folder. If the transaction is rolled back, they will be moved back into place. If the transaction is committed, these files are automatically cleaned up using the temp system.

All operations that create files will automatically create any intermediate directories in the path to the file if they do not already exist.

All operations that create files will fail if the destination already exists.

Methods

impl<'a> Transaction<'a>

fn new(prefix: InstallPrefix, temp_cfg: &'a Cfg, notify_handler: NotifyHandler<'a>) -> Self

fn commit(self)

Commit must be called for all successful transactions. If not called the transaction will be rolled back on drop.

fn add_file(&mut self, component: &str, relpath: PathBuf) -> Result<File>

Add a file at a relative path to the install prefix. Returns a File that may be used to subsequently write the contents.

fn copy_file(&mut self, component: &str, relpath: PathBuf, src: &Path) -> Result<()>

Copy a file to a relative path of the install prefix.

fn copy_dir(&mut self, component: &str, relpath: PathBuf, src: &Path) -> Result<()>

Recursively copy a directory to a relative path of the install prefix.

fn remove_file(&mut self, component: &str, relpath: PathBuf) -> Result<()>

Remove a file from a relative path to the install prefix.

fn remove_dir(&mut self, component: &str, relpath: PathBuf) -> Result<()>

Recursively remove a directory from a relative path of the install prefix.

fn write_file(&mut self, component: &str, relpath: PathBuf, content: String) -> Result<()>

Create a new file with string contents at a relative path to the install prefix.

fn modify_file(&mut self, relpath: PathBuf) -> Result<()>

If the file exists back it up for rollback, otherwise ensure that the path to it exists so that subsequent calls to File::create will succeed.

This is used for arbitrarily manipulating a file.

fn temp(&self) -> &'a Cfg

fn notify_handler(&self) -> NotifyHandler<'a>