Struct rustup_dist::manifestation::Manifestation

pub struct Manifestation {
    // some fields omitted
}

Methods

impl Manifestation

fn open(prefix: InstallPrefix, triple: TargetTriple) -> Result<Self>

Open the install prefix for updates from a distribution channel. The install prefix directory does not need to exist; it will be created as needed. If there's an existing install then the rust-install installation format will be verified. A bad installer version is the only reason this will fail.

fn update(&self, new_manifest: &Manifest, changes: Changes, temp_cfg: &Cfg, notify_handler: NotifyHandler) -> Result<UpdateStatus>

Install or update from a given channel manifest, while selecting extension components to add or remove.

update takes a manifest describing a release of Rust (which may be either a freshly-downloaded one, or the same one used for the previous install), as well as lists off extension components to add and remove. From that it schedules a list of components to uninstall and to uninstall to bring the installation up to date. It downloads the components' packages. Then in a Transaction uninstalls old packages and installs new packages, writes the distribution manifest to "rustlib/rustup-dist.toml" and a configuration containing the component name-target pairs to "rustlib/rustup-config.toml".

fn uninstall(&self, temp_cfg: &Cfg, notify_handler: NotifyHandler) -> Result<()>

fn read_config(&self) -> Result<Option<Config>>

fn load_manifest(&self) -> Result<Option<Manifest>>

fn update_v1(&self, new_manifest: &[String], update_hash: Option<&Path>, temp_cfg: &Cfg, notify_handler: NotifyHandler) -> Result<Option<String>>

Installation using the legacy v1 manifest format